Only use 1GB pages when supported.

+++ src/hip.cpp
@@ -35,9 +35,9 @@ void Hip::build (mword addr)
 
     h->signature  = 0x41564f4e;
     h->cpu_offs   = reinterpret_cast<mword>(h->cpu_desc) - reinterpret_cast<mword>(h);
-    h->cpu_size   = sizeof (Hip_cpu);
+    h->cpu_size   = static_cast<uint16>(sizeof (Hip_cpu));
     h->mem_offs   = reinterpret_cast<mword>(h->mem_desc) - reinterpret_cast<mword>(h);
-    h->mem_size   = sizeof (Hip_mem);
+    h->mem_size   = static_cast<uint16>(sizeof (Hip_mem));
     h->api_flg    = FEAT_VMX | FEAT_SVM;
     h->api_ver    = CFG_VER;
     h->sel_num    = Space_obj::caps;
